+++ src/components/PaginationButtons/PaginationButtons.styled.ts
@@ -1,0 +1,7 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+
+})
+
+export default styles
